feat: sim now functions

+++ src/main/java/org/team1540/robot2024/Constants.java
@@ -129,9 +129,9 @@ public final class Constants {
         public static final double KI = 0;
         public static final double KD = 0.1;
         public static final double KG = 0;
-        public static final int CRUISE_VELOCITY_MPS = 2;
-        public static final int MAXIMUM_ACCELERATION_MPS2 = 20;
-        public static final int JERK_MP3 = 40;
+        public static final double CRUISE_VELOCITY_MPS = 2;
+        public static final double MAXIMUM_ACCELERATION_MPS2 = 20;
+        public static final double JERK_MP3 = 40;
         public static final double SPROCKET_CIRCUMFERENCE = .044 * Math.PI;
         public static final double MOTOR_ROTS_TO_METERS = GEAR_RATIO * SPROCKET_CIRCUMFERENCE;
         public static final double ERROR_TOLERANCE = 0.03;
